third_party/blink/web_tests/platform/generic/http/tests/devtools/console/console-copy-truncated-text-expected.txt

Tests that console copies truncated text in messages properly.

Message count: 8
Long url has max length: 150, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z…z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Running: testSelectWithinTruncatedUrl

Making selection: 1, 0, 1, 75
Selection length: 75, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z

Making selection: 1, 0, 1, 76
Selection length: 84, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789

Making selection: 1, 0, 1, 150
Selection length: 158, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Making selection: 1, 75, 1, 76
Selection length: 9, text: 123456789

Making selection: 1, 75, 1, 150
Selection length: 83, text: 123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Making selection: 1, 76, 1, 150
Selection length: 74, text: z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Running: testSelectAcrossMultipleMessages

Making selection: 1, 0, 2, 11
Selection length: 175, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
VM:5 www.bar.com

Making selection: 1, 75, 2, 11
Selection length: 100, text: 123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
VM:5 www.bar.com

Making selection: 1, 76, 2, 11
Selection length: 91, text: z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
VM:5 www.bar.com

Running: testSelectAcrossMultipleMessagesWithTruncatedUrls

Making selection: 1, 0, 3, 75
Selection length: 256, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
VM:5 www.bar.com
VM:6 www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z

Making selection: 1, 0, 3, 76
Selection length: 265, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
VM:5 www.bar.com
VM:6 www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789

Making selection: 1, 0, 3, 150
Selection length: 339, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
VM:5 www.bar.com
VM:6 www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Running: testSelectWithinMessageWithMultipleTruncatedUrls

Making selection: 4, 0, 4, 75
Selection length: 75, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z

Making selection: 4, 0, 4, 76
Selection length: 84, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789

Making selection: 4, 0, 4, 163
Selection length: 171, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com 

Making selection: 4, 0, 4, 238
Selection length: 246, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z

Making selection: 4, 0, 4, 239
Selection length: 255, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789

Making selection: 4, 0, 4, 313
Selection length: 329, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Making selection: 4, 75, 4, 76
Selection length: 9, text: 123456789

Making selection: 4, 75, 4, 163
Selection length: 96, text: 123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com 

Making selection: 4, 75, 4, 238
Selection length: 171, text: 123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z

Making selection: 4, 75, 4, 239
Selection length: 180, text: 123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789

Making selection: 4, 75, 4, 313
Selection length: 254, text: 123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Making selection: 4, 76, 4, 163
Selection length: 87, text: z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com 

Making selection: 4, 76, 4, 238
Selection length: 162, text: z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z

Making selection: 4, 76, 4, 239
Selection length: 171, text: z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789

Making selection: 4, 76, 4, 313
Selection length: 245, text: z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Making selection: 4, 163, 4, 238
Selection length: 75, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z

Making selection: 4, 163, 4, 239
Selection length: 84, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789

Making selection: 4, 163, 4, 313
Selection length: 158, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Making selection: 4, 238, 4, 239
Selection length: 9, text: 123456789

Making selection: 4, 238, 4, 313
Selection length: 83, text: 123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Making selection: 4, 239, 4, 313
Selection length: 74, text: z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Running: testSelectWithinShortUrlWithHashes

Making selection: 5, 0, 5, 14
Selection length: 26, text: www.01234567890123456789zf

Making selection: 5, 0, 5, 28
Selection length: 52, text: www.01234567890123456789zfoobarz01234567890123456789

Making selection: 5, 14, 5, 28
Selection length: 26, text: oobarz01234567890123456789

Running: testSelectWithinUrlWithHashes

Making selection: 6, 0, 6, 63
Selection length: 75, text: www.01234567890123456789zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 6, 0, 6, 64
Selection length: 120, text: www.01234567890123456789zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 6, 0, 6, 126
Selection length: 194, text: www.01234567890123456789zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz01234567890123456789

Making selection: 6, 63, 6, 64
Selection length: 45, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 6, 63, 6, 126
Selection length: 119, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz01234567890123456789

Making selection: 6, 64, 6, 126
Selection length: 74, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz01234567890123456789

Running: testSelectWithinHighlightedUrlBeginning
Searching for text: www.
Highlighted 18 matches

Making selection: 7, 0, 7, 75
Selection length: 75, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 7, 0, 7, 76
Selection length: 134, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 7, 0, 7, 150
Selection length: 208, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Making selection: 7, 75, 7, 76
Selection length: 59, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 7, 75, 7, 150
Selection length: 133, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Making selection: 7, 76, 7, 150
Selection length: 74, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Running: testSelectWithinHighlightedUrlMiddle
Searching for text: zzzzz
Highlighted 118 matches

Making selection: 7, 0, 7, 75
Selection length: 75, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 7, 0, 7, 76
Selection length: 134, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 7, 0, 7, 150
Selection length: 208, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Making selection: 7, 75, 7, 76
Selection length: 59, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 7, 75, 7, 150
Selection length: 133, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Making selection: 7, 76, 7, 150
Selection length: 74, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Running: testSelectWithinHighlightedUrlEnd
Searching for text: .com
Highlighted 14 matches

Making selection: 7, 0, 7, 75
Selection length: 75, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 7, 0, 7, 76
Selection length: 134, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 7, 0, 7, 150
Selection length: 208, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Making selection: 7, 75, 7, 76
Selection length: 59, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 7, 75, 7, 150
Selection length: 133, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Making selection: 7, 76, 7, 150
Selection length: 74, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com
